chore: dapr supports the default AppPort obtained from ApplicationUrl by default, automatically assigns the grpc and http ports of dapr and automatically supplements environment variables
Supports automatic restart after the dapr process is closed, and the configuration after restart is consistent with the last successful startup configuration

Supports automatic restart of the dapr process after configuration update. Currently, there are no restrictions on HttpPort and GrpcPort. However, if the configuration update changes HttpPort and GrpcPort, the port obtained by DaprClient will be inconsistent with the actual running one. It needs to be adjusted later here.

+++ src/Development/MASA.Utils.Development.Dapr/Process/SystemProcess.cs
@@ -9,13 +9,15 @@ public class SystemProcess : IProcess
         _process = process;
     }
 
-    public int PId => _process.Id;
+    public int PId => IsRun ? _process.Id : 0;
 
-    public string Name => _process.ProcessName;
+    public string Name => IsRun ? _process.ProcessName : string.Empty;
+
+    public bool IsRun => !_process.HasExited;
 
     public void Kill()
     {
-        if (!_process.HasExited)
+        if (IsRun)
             _process.Kill();
     }
 
@@ -27,11 +29,7 @@ public class SystemProcess : IProcess
         {
             _process.WaitForExit(milliseconds.Value);
         }
-        else if (!_process.HasExited)
-        {
-            _process.WaitForExit();
-        }
-        else
+        else if (IsRun)
         {
             _process.Kill();
         }
